Break out processor functions
